% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSQLtable.R
\name{readSQLtable}
\alias{readSQLtable}
\title{Read SQL from Databricks}
\usage{
readSQLtable(tableName, token, query = NULL, hostName = NULL,
  httpPath = NULL)
}
\arguments{
\item{tableName}{Name of the table to read.}

\item{token}{DataBricks token. IT IS PERSONAL TO YOU.}

\item{query}{SQL query. Default: "SELECT * FROM tableName"}

\item{hostName}{Server host name. Default: "db.ref.core.data.digital.nhs.uk"}

\item{httpPath}{HTTP path to the Databricks cluster. Default: "sql/protocolv1/o/0/rstudio-pilot"}
}
\value{
The SQL table.
}
\description{
Reading an SQL table from Databricks assumes that the user already has Databricks credentials
and access rights to a particular Databricks cluster.
}
\examples{

}
